* commit 'ad68bbf1e5848b2482b1f59417a5eaf1b83037a0': (23 commits)
  KERNEL-2362 focuspoint表分子表导入性能优化 1、无脑异常性能堪忧，调整为index判断，且-1涵盖99%的场景。
  KERNEL-2314 分析任务计算前做块的预处理
  KERNEL-2255 swift精确计算适配，尝试提高导入完成后初次计算的效率。 1、完善接入精确计算 2、导入任务的cause填对应分析任务的taskContent内容，方便关联 3、ASM类命名调整，并且会将计算用到的，缓存中不存在，数据库中存在的块放入缓存
  DEC-11023【swift】导入任务进行时，分析日志里出现大量报错 1、try append 还有其他异常，先重新append20次再说。
  DEC-11100 【埋点分析】多服务任务并行处理方案研究 1、根据任务数量进行均衡的策略实现。 2、kafkaconsumer监听springboot服务器ready事件后再起动。
  DEC-11100 【埋点分析】多服务任务并行处理方案研究
  无jira任务 更改允许collate操作的任务配置项
  DEC-11074 【swift】导入和分析分节点执行，分析任务日志中出现大量块的报错信息 暂时修改成只有用于分析的那台机器可以进行Collate
  DEC-11022 【swift】分析节点不能实时刷新导入节点新增的块，导致分析结果表有异常 更改每次分析都重新读数据库配置 为每次根据最新的分析信息更新缓存
  DEC-11022 【swift】分析节点不能实时刷新导入节点新增的块，导致分析结果表有异常
  DEC-11023【swift】导入任务进行时，分析日志里出现大量报错 1、先index完，再save location
  无jira任务 修复分析任务不开始
  KERNEL-1793 swift服务端tcp完善+查询接口定义实现 1、本地客户端测试代码; 2、流程数据保存处理; 3、查询接口定义,查询逻辑实现;
  KERNEL-1635 collate完善 1、segments引用bug
  DEC-10870【埋点分析】customer_base_info表中数据缺失 1、v3 customer_base_info表结果完善
  KERNEL-1619 任务队列可以获取指定类型的任务
  REPORT-23475 swiftjdbc默认聚合方式修改 Fix String默认单值  数值默认COUNT
  KERNEL-1635 collate完善 1、collate块数过多，任务无法入库 2、collate分批执行，每次10块，避免某一块问题导致所有快都无法collate 3、collate、import、analysis不能同时执行。
  feature:适配fineio读写，支持long的pos
  无jira任务 提升版本，git警告了
  ...

+++ swift-base/swift-base-io/src/main/java/com/fr/swift/cube/io/impl/fineio/input/LongFineIoReader.java
@@ -27,6 +27,6 @@ public class LongFineIoReader extends BaseFineIoReader<LongBuf> implements LongR
 
     @Override
     public long get(long pos) {
-        return FineIOAccessor.INSTANCE.getLong(readFile, (int) pos);
+        return FineIOAccessor.INSTANCE.getLong(readFile, pos);
     }
 }
