+++ swift-base/swift-base-io/src/main/java/com/fr/swift/cube/io/impl/fineio/output/IntFineIoWriter.java
@@ -35,9 +35,9 @@ public class IntFineIoWriter extends BaseFineIoWriter<IntBuf> implements IntWrit
     @Override
     public void put(long pos, int val) {
         if (isOverwrite) {
-            FineIOAccessor.INSTANCE.put(writeFile, (int) pos, val);
+            FineIOAccessor.INSTANCE.put(writeFile, pos, val);
         } else {
-            FineIOAccessor.INSTANCE.put(appendFile, (int) pos, val);
+            FineIOAccessor.INSTANCE.put(appendFile, pos, val);
         }
     }
 }
